refa: racingGame을 이용하여 비즈니스로직 분리
