--- src/main/java/racingcar/domain/car/CarRegisterer.java
@@ -1,24 +1,0 @@
-package racingcar.domain.car;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class CarRegisterer {
-
-    private final List<String> carNames;
-
-    public CarRegisterer() {
-        this.carNames = new ArrayList<>();
-    }
-
-    public void registerCar(String carName) {
-        carNames.add(carName);
-    }
-
-    public Cars prepareCars() {
-        return new Cars(carNames.stream()
-                .map(carName -> new Car(new Name(carName)))
-                .collect(Collectors.toUnmodifiableList()));
-    }
-}
